Add backend contract adapter

+++ api_detection/contracts.py
@@ -54,7 +54,7 @@ class RequestInfo:
 @dataclass(frozen=True)
 class ResponseInfo:
     status_code: int
-    latency_ms: int | None = None
+    latency_ms: float | None = None
 
 
 @dataclass(frozen=True)
